5.07 Refactor DetailFragment into a new file
